feat : 질문 조회 시 조회수 증가 구현

+++ server/src/main/java/preproject/stackoverflow/question/entity/Question.java
@@ -29,7 +29,7 @@ public class Question {
     @Column(nullable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
     @Column(nullable = false)
-    private Integer view = 0;
+    private Long view = 0L;
 
     // 질문 상태 : 등록, 답변 완료, 삭제
     @Enumerated(EnumType.STRING)
